baseline logic
Added first-page logic. Needed modifications and the revision in which
they will undergo these necessary modifications is noted in TODO
comments. Some BASIC ones are already complete, but I'm not uploading
them until I get them working in baseline format, and I'm not uploading
the classes which are not relevant to workfloor logic.