9 GS changes.
26-28:
Trimmed a bit to date.

106-201:
Moved the incredibly stupid algorithm to Controller. Now at least it's a little less stupid. Still stupid. Rewrite. Now.

228-230:
Added some more inline settings until the admin control panel settings are up.

373:
Changed it to be a bit more specific, what with there being two functions now.

468-620:
YES. DIE.

716, 834-844:
Added JS property to hold Javascripts that need loading, as well as a function to approach it.

911-912:
Added core url and trimmed core url global vars.

1053-1063:
Merged the primary display name and the current display name to be in the same to-be-loaded array. Then the front-end isn't cringeworthy.

1068-1082:
Now made it so that the primary name is first.

Why would I do that? The last-used name should be first. Stupid, stupid, stupid!!